Adapt c nondet symbol factory for use in function body generation
We essentially make two changes to the c nondet symbol factory:
- A parameter assign_const is added to gen_nondet_init(). When false, an object
  of const type is not nondet assigned. This will be used in the function body
  generation feature to not havoc const arguments (e.g., the targets of pointers
  to const int)
- A parameter lifetime (an enum of type lifetimet) is added to the constructor
  of symbol_factoryt. It indicates whether it uses lifetimet::AUTOMATIC_LOCAL,
  lifetimet::STATIC_GLOBAL, or lifetimet::DYNAMIC when allocating new objects
  with allocate_object(). Previously this was hardcoded to behave like
  lifetime::AUTOMATIC_LOCAL. However, in the function body generation we need to
  add code to a function to allocate dynamic objects. This is because the
  function may be called multiple times, hence using local or global variables
  does not work.
